requirements.txt
python-dotenv==1.0.0
pydantic==2.12.0
pydantic-settings==2.12.0
flask-session<0.6.0
apache-airflow==2.10.2
apache-airflow-providers-redis==3.5.0
apache-airflow-providers-postgres==5.10.0
aiohttp==3.9.1
requests==2.32.0
pybit==5.7.0
pycryptodome==3.19.0
ccxt==4.2.25
pandas==2.1.4
numpy==1.26.2
redis==4.5.2
psycopg2-binary==2.9.9
asyncpg>=0.29.0
sqlalchemy==1.4.36
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0
python-telegram-bot==21.9
pytest==7.4.3
pytest-asyncio==0.23.2